微修正

+++ backend/internal/model/payment.go
@@ -11,7 +11,7 @@ type Payment struct {
 	OrderID               uint           `gorm:"not null;uniqueIndex" json:"order_id"`
 	StripePaymentIntentID string         `gorm:"size:255;index" json:"stripe_payment_intent_id"`
 	StripePaymentMethodID string         `gorm:"size:255" json:"stripe_payment_method_id,omitempty"`
-	Amount                float64        `gorm:"not null" json:"amount"`
+	Amount                int            `gorm:"not null" json:"amount"` // 最小通貨単位（日本円の場合は円単位）
 	Currency              string         `gorm:"default:'jpy'" json:"currency"`
 	Status                string         `gorm:"default:'pending'" json:"status"` // pending, succeeded, failed, canceled
 	CreatedAt             time.Time      `json:"created_at"`
